% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{freq_table}
\usage{
freq_table(data_frame_string, column_string, sort_by_count = FALSE,
  total_row = TRUE)
}
\arguments{
\item{data_frame_string}{a string of the name of a data frame}

\item{column_string}{a string of the column to be aggregated}

\item{sort_by_count}{a boolean value that determines if the output will be sorted by count or name}

\item{total_row}{a boolean value that determines if the output will have a summary row appended}
}
\value{
a data_frame containing the counts and percentages of each value from the provided data
}
\description{
freq_table returns a frequency table with counts and percentages of values
  from the provided data frame and column.
}
\examples{
tbl <- data.frame(numbers = sample(1:10, 200, replace = TRUE),
  letters = sample(letters, 200, replace = TRUE),
  dates = sample(seq(as.Date('1999/10/01'), as.Date('2000/01/01'), by="day"), 200, replace = TRUE),
  logicals = sample(c(TRUE, FALSE), 200, replace = TRUE),
  stringsAsFactors = FALSE)

freq_table('tbl', 'numbers')
View(freq_table('tbl', 'letters'))
View(freq_table('tbl', 'letters', sort_by_count = TRUE, total_row = FALSE))
}
