% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table2.R
\name{freq_table2}
\alias{freq_table2}
\title{freq_table2}
\usage{
freq_table2(df_string, col1_string, col2_string)
}
\arguments{
\item{df_string}{a string of the name of a data frame}

\item{col1_string}{a string of the column name to be aggregated at the higher level.}

\item{col2_string}{a string of the column name to be aggregated within col1_string.}
}
\value{
returns a list containing frequency tables split by col1_string with counts and rates of
        col2_string.
}
\description{
freq_table2 takes two columns from a data frame and returns a list containing frequency tables
   with counts and percentages of col2 within col1.
   It answers the question, what percent of col1 is col2.
}
\details{
For example, if col1 was gender and col2 was ethnicity you would get the count of rate of
  males that are asian, african american, hispanic, etc.

If col1 was ethnicity and col2 was gender you would get the count and rate of asians that are
  male and female.
}
\examples{
df <- data.frame(gender = sample(c('m','f'), 200, replace = TRUE),
                 ethnicity = sample(c('african american', 'asian', 'caucasian', 'hispanic', 'other'),
                                   200, replace = TRUE))
freq_table2('df', 'gender', 'ethnicity')
gender_by_ethnicity <- freq_table2('df', 'gender', 'ethnicity')
gender_by_ethnictiy$m
freq_table2('df', 'gender', 'ethnicity')$m
freq_table2('df', 'ethnicity', 'gender')
ethnicity_by_gender <- freq_table2('df', 'ethnicity', 'gender')
ethnicity_by_gender$asian
}

