% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{freq_table}
\usage{
freq_table(data_frame_string, column_string)
}
\arguments{
\item{data_frame_string}{a string of the name of a data frame}

\item{column_string}{a string of the column to be aggregated}
}
\value{
a tibble containing the counts and percentages of each value from the provided data
}
\description{
freq_table returns a frequency table with counts and percentages of values
  from the provided data frame and column.
}
\examples{
tbl <- data_frame(numbers = sample(1:10, 200, replace = TRUE),
                  letters = sample(letters, 200, replace = TRUE))
freq_table('tbl', 'numbers')
View(freq_table('tbl', 'letters'))
}

